To build new os systems (#3)
The obvious one is 'because you can'.
You may have new hardware that you do not want to distribute code for, GPL code would not be available for that purpose but BSD would, it may be possible that these do not meet your requirements because of something unforseen.

If there are no other reasons apart from learning how the entire structure gets up and running, I would recommend using a very base linux distribution Welcome to Linux From Scratch! to learn how an OS can be put together and how an it will interact with the surrounding hardware. LFS and its forums will take you by the hand on your learning journey